agentic-webscraper/requirements.txt
requests
python-dotenv
beautifulsoup4
ddgs
langchain-openai
langchain-core
